meshrr/requirements.txt
bcrypt==3.2.0
cffi==1.14.4
cryptography==3.2.1
dnspython==2.0.0
importlib-resources==3.3.0
Jinja2==2.11.2
junos-eznc==2.5.4
lxml==4.6.1
MarkupSafe==1.1.1
ncclient==0.6.9
netaddr==0.8.0
paramiko==2.7.2
pip==9.0.1
pycparser==2.20
PyNaCl==1.4.0
pyparsing==2.4.7
pyserial==3.5
python-dotenv==0.15.0
PyYAML==5.3.1
scp==0.13.3
setuptools==39.0.1
six==1.15.0
transitions==0.8.5
wheel==0.35.1
yamlordereddictloader==0.4.0
zipp==3.4.0